Day 10 code fix

+++ Day10/calculator.py
@@ -23,24 +23,20 @@ def subtract(a , b):
 
 operation_dict = { "+" : add , "-" : subtract , "*" : multiply , "/" : divide}
 state = 'n'
-# a = int( input("What's the first number?: ") )
-# for symbol in operation_dict:
-#     print(symbol)
-# choosen_operation = input("Pick an operation: ")
-# b = int( input("What's the next number?: ") )
+
 
 def calculator():
     
     print(logo)
     should_continue = True  
-    a = int( input("What's the first number?: ") )
+    a = float( input("What's the first number?: ") )
     for symbol in operation_dict:
             print(symbol)
 
     while should_continue:
 
         choosen_operation = input("Pick an operation: ")
-        b = int( input("What's the next number?: ") )
+        b = float( input("What's the next number?: ") )
         result = operation_dict[choosen_operation](a , b)
 
         print(f"{a} {choosen_operation} {b} = {result}")
@@ -52,8 +48,6 @@ def calculator():
             should_continue = False
             os.system('cls')
             calculator()
-
-
 
 
 calculator()
